Add snyk monitoring
Description
===========

This patch adds snyk monitoring to the build pipeline.
It will hook itself into the check and publish stages.

The patch also sets a dependency helper plugin
`net.wooga.cve-dependency-resolution` which applies overrides
for dependencies with know fixes for security issues.

Along with the introduction of snyk I also upgraded/removed some
depdencies. Coveralls produces a huge load of errors even with the
latest version so I decided to remove it since we want/are moving
to sonarqube (It is unknown at this time if this dependency is actually better
or not).

Changes
=======

* ![ADD] `snyk` monitoring
* ![ADD] `net.wooga.snyk-wdk-java` snyk convention plugin
* ![ADD] `net.wogoa.cve-dependency-resolution` plugin
* ![REMOVE] coveralls plugin
* ![UPDATE] `org.codehaus.groovy:groovy-all` to version `2.5.16`
* ![UPDATE] `net.lingala.zip4j:zip4j` to version `2.10.0`
* ![UPDATE] `commons-io:commons-io` to version `[2.5,3)`